My Sites: Improve speed of switch from Reader to My Sites
When user switches between the Reader and My Sites there is a fairly long pause, this commit tries to reduce this pause duration.

It achieves it by turning MySitesSidebar to be pure component and thus eliminates wasted render calls when the data for this component  do no change.

Since this component consume SitesList which mutate in-place `SitesObserver` component was added to handle changes in SitesList and propogate them in way that will update the PureComponent, specifically by creating new instance which uses SitesList as prototype but will have different memory address and thus shallow compare will detect as changed.